+++ src/RecoveryServices/RecoveryServices/ChangeLog.md
@@ -22,6 +22,8 @@
 * Added Azure Site Recovery VmwareToAzure properties DR monitoring
 * Azure Backup added support for retrying policy update for failed items.
 * Azure Backup Added support for disk exclusion settings during backup and restore.
+* Azure Backup Added Support for Restoring Multiple files/folders in AzureFileShare
+* Azure Backup Added support for User-specified Resourcegroup support while updating IaasVM Policy
 
 ## Version 2.7.0
 * Added Support for Restore-as-files for SQL Databases.